remove banners and fix callouts (#82)
There were 2 banners left and both of them had old wordmarks so I just
deleted them. If it's really important we add them back (like if we
needed it there for a partnership with Zapier or something) we need to
make new images with our new wordmark. I also removed an empty codeblock
and converted all `<Info>` blocks to `<Note>` to match mintlify
convention.